Implemented a basic web UI to configure upload limits per user
This still lacks the option to set a default. It also still lacks an
option to set a maximum concurrent uploads. It would probably also make
sense to implement a download limiter as well, for the actual
downloading of files of course. Such a limit would make no sense for the
general web UI.

But most importantly, the admin panel code is very nasty atm and should
be fixed up first. The separate pieces should just get their own handler
each, rather than hacky build on top of 1 like it is right now. For this
'prototype' for the upload limit part it just made a bit more sense to
stick with it. It however got even more nasty in the process.

+++ pkg/http/helper/limiter/middleware.go
@@ -37,7 +37,7 @@ func (m *middleWare) Serve(user *models.User, w http.ResponseWriter, r *http.Req
 	logrus.Debugf("Applying the following rate limits: %#v", limit)
 
 	// otherwise we override the body with the freshly fetched settings
-	r.Body = NewReader(r.Body, limit.RateLimit, limit.Burst)
+	r.Body = NewReader(r.Body, float64(limit.RateLimit), limit.Burst)
 
 	// and we pass the request with the adjusted body along to the wrapped handler
 	m.handler.Serve(user, w, r)
